requirements.txt
torch==1.8.1+cu111
attrdict
distlib
gym
imageio
matplotlib
numpy
opencv-python
Pillow
pygame
PyYAML
rl-games
scipy
tensorboardX
termcolor
timm
tqdm
